+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - [#227](https://github.com/deviceinsight/kafkactl/issues/227) Incorrect handling of Base64-encoded values when producing from JSON
 - [#228](https://github.com/deviceinsight/kafkactl/issues/228) Fix parsing version of gcloud kubectl.
+- [#217](https://github.com/deviceinsight/kafkactl/issues/217) Allow reset of consumer-group in dead state
 
 ## 5.4.0 - 2024-11-28
 ### Added